refactor(eval): extract report formatters
